move world processor function declarations to world.d.ts
